Rename v1 issue card
