+++ src/applications/disability-benefits/996/tests/components/IssueCard.unit.spec.jsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { shallow, mount } from 'enzyme';
 import sinon from 'sinon';
 
-import { IssueCardContent, IssueCard } from '../../components/IssueCard';
+import { IssueCardContent, IssueCard } from '../../components/IssueCardV1';
 import { SELECTED } from '../../constants';
 
 const getContestableIssue = (id, selected) => ({
